Create utils file for MapView component
